доработал запуск в Docker сделал Admin: Пользователи        API для работы с пользователями

+++ main-service/src/main/java/ru/practicum/dtos/request/ParticipationRequestDto.java
@@ -18,17 +18,17 @@ import static ru.practicum.dtos.utils.DateTimeFormatter.FORMAT;
 @Builder
 public class ParticipationRequestDto {
 
-    Integer id;
+    Long id;
 
     @NotNull
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = FORMAT)
     LocalDateTime created;
 
     @NotNull
-    Integer event; // Идентификатор события
+    Long event; // Идентификатор события
 
     @NotNull
-    Integer requester; // Идентификатор пользователя
+    Long requester; // Идентификатор пользователя
 
     @NotNull
     Status status;
